requirements.txt
PySide6>=6.6
PySide6-Essentials>=6.6
python-dotenv>=1.0.1
mysql-connector-python>=9.0.0
bcrypt>=4.2.0
reportlab>=4.2.2
